digital pots functionality added, need to test
